Sync master to 2.3.0 branch

+++ samples/client/petstore/tizen/src/PetManager.h
@@ -1,19 +1,3 @@
-/*
- * Copyright (c) 2016 Samsung Electronics Co., Ltd All Rights Reserved
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
 #ifndef _PetManager_H_
 #define _PetManager_H_
 
@@ -21,8 +5,8 @@
 #include <cstring>
 #include <list>
 #include <glib.h>
-#include "Pet.h"
 #include "ApiResponse.h"
+#include "Pet.h"
 #include "Error.h"
 
 namespace Tizen{
@@ -67,7 +51,7 @@ bool addPetAsync(char * accessToken,
  * \param userData The user data to be passed to the callback function.
  */
 bool deletePetSync(char * accessToken,
-	long petId, std::string apiKey, 
+	long long petId, std::string apiKey, 
 	
 	void(* handler)(Error, void* ) , void* userData);
 
@@ -81,7 +65,7 @@ bool deletePetSync(char * accessToken,
  * \param userData The user data to be passed to the callback function.
  */
 bool deletePetAsync(char * accessToken,
-	long petId, std::string apiKey, 
+	long long petId, std::string apiKey, 
 	
 	void(* handler)(Error, void* ) , void* userData);
 /*! \brief Finds Pets by status. *Synchronous*
@@ -143,7 +127,7 @@ bool findPetsByTagsAsync(char * accessToken,
  * \param userData The user data to be passed to the callback function.
  */
 bool getPetByIdSync(char * accessToken,
-	long petId, 
+	long long petId, 
 	void(* handler)(Pet, Error, void* )
 	, void* userData);
 
@@ -156,7 +140,7 @@ bool getPetByIdSync(char * accessToken,
  * \param userData The user data to be passed to the callback function.
  */
 bool getPetByIdAsync(char * accessToken,
-	long petId, 
+	long long petId, 
 	void(* handler)(Pet, Error, void* )
 	, void* userData);
 /*! \brief Update an existing pet. *Synchronous*
@@ -195,7 +179,7 @@ bool updatePetAsync(char * accessToken,
  * \param userData The user data to be passed to the callback function.
  */
 bool updatePetWithFormSync(char * accessToken,
-	long petId, std::string name, std::string status, 
+	long long petId, std::string name, std::string status, 
 	
 	void(* handler)(Error, void* ) , void* userData);
 
@@ -210,7 +194,7 @@ bool updatePetWithFormSync(char * accessToken,
  * \param userData The user data to be passed to the callback function.
  */
 bool updatePetWithFormAsync(char * accessToken,
-	long petId, std::string name, std::string status, 
+	long long petId, std::string name, std::string status, 
 	
 	void(* handler)(Error, void* ) , void* userData);
 /*! \brief uploads an image. *Synchronous*
@@ -224,7 +208,7 @@ bool updatePetWithFormAsync(char * accessToken,
  * \param userData The user data to be passed to the callback function.
  */
 bool uploadFileSync(char * accessToken,
-	long petId, std::string additionalMetadata, std::string file, 
+	long long petId, std::string additionalMetadata, std::string file, 
 	void(* handler)(ApiResponse, Error, void* )
 	, void* userData);
 
@@ -239,7 +223,7 @@ bool uploadFileSync(char * accessToken,
  * \param userData The user data to be passed to the callback function.
  */
 bool uploadFileAsync(char * accessToken,
-	long petId, std::string additionalMetadata, std::string file, 
+	long long petId, std::string additionalMetadata, std::string file, 
 	void(* handler)(ApiResponse, Error, void* )
 	, void* userData);
